Calculate the uuid from pid, pver, prev and serial
If no UUID is supplied we calculate the uuid from the pid, pver, prev
and serial. All these values are concatenated and an md5sum is
calculated over these values. The uuid is then generated from the
md5sum.

The RFC4122 describes in section 4.4 a way to create UUIDs from random data.
http://tools.ietf.org/html/rfc4122#section-4.4
We use the same algorithem to calculate our UUIDs. But we use the md5sum
of the pid, pver, prev and serial instead of random data.

This will ensure we get a fairly unique UUID for each device.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>