Replace u16 and u32 parser with generic implementation
Use the num crate to implement a generic implementation for the u16 and
u32 parser. The num::Unsigned trait can be used to limit the generic to
only apply to numbers which are unsigned. And thus implement
from_str_radix().

Also we limit T further to only apply to types where FromstrRadixErr is
ParseIntError. Thus leavs our implementation only for unsigned integers.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>