Use array to define gpiobank(s)
Currently only the first bank can be configured by the HAT EEPROM. But
the BCM28xx has more gpio banks. If at any point more banks could be
configured the json format would need to change. So we now use an array
of banks even if we don't allow to configure more then the first bank.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>