Add eepmake example
This example should be compatible with the configuration file formati[1]
used by the rpi eepmake utility[2]. It is not intended as a bug2bug
compatible implemenation. It is intended to check if the rpi-eep crate
is creating correct images. And as an example on how to use the rpi-eep
crate.

[1] https://github.com/raspberrypi/hats/blob/master/eepromutils/eeprom_settings.txt
[2] https://github.com/raspberrypi/hats/blob/master/eepromutils/eepmake.c

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>