rpi-hat-eep: gpio_map: Fix calc. bank_drive value
The bank_drive field of the GPIO map is a byte. The lower four bits are
used for the drive (strength), bits 4 and 5 are used for the slew (rate)
and the upper two bits are used for the hysteresis.

The mask for the values are wrong. The are done on the input values
before the shift. And so the slew rate and hysteresis are always 0.

Use the correct mask to fix this.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>