revpi-hat-eep: Implement TryFrom<GpioBank> for EepAtomGpioMapData
To convert GpioBank objects into EepAtomGpioMapData object we can use
the TryFrom trait. For this implementation to work we need the Copy
trait satisfied for the enum members of the GpioBank struct. And as we
are using only simple enums which resolve to small integers the Copy
trait is efficient as well.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>