rpi-hat-eep: Replace push() with set() for EepAtomGpioMapData
The push() function of the EepAtomGpioMapData struct has some drawbacks.
The user is responsible to push the full list of gpios. Even if the user
doesn't want to configure any gpio he needs to cal push() repeatedly.

This commit replaces the push() function with a set() function. The full
gpio list is now created when the new() function is called. And the user
can canfigure any gpio by caling the set() function with the gpio
number.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>