revpi-eep: Add a --export option to save a full json config
With the --export option a full json config is saved to the given file.
The full config includes the optional attributes serial, edate and mac.
In the future it will be possible to recreate an eep from a full config.
Currently the args --serial, --edate and --mac are still mandatory and
will overwrite any optional values from a given config.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>