hat-eeprom-format: Update the prod. version  description
The description of how to interpret the version is wrong. If the prod
version as u16 is the version * 100, it would not be possible to see x.y
as major and minor version. And it would not be possible to have a 1.10
after a 1.9 version. Thus the description is now changed to reflect the
correct interpretation of the product version.
For this no change to the revpi-hat-eep crate or the revpi-eep program
is necessary.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>