rpi-hat-eep: Refactor the atom order validation code
The order of the atoms is defined in section "EEPROM Structure" here:
https://github.com/raspberrypi/hats/blob/master/eeprom-format.md#eeprom-structure

The current code uses the position of the atom which is added to ensure
the correct order. This is working, but makes it hard to extend this
further. And with the gpio1 bank support this will get really messy if
not impossible.

This patch just uses the type of the last atom to check if the new atom
is valid.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>