feat: allow including definition
The HAT EEPROM definitions can be very similar to each other if the same
device has different definitions that are slightly similar. To ease the
maintenance of these files it is now possible to include other files in
a definition.

These template files can only define the "gpiobanks" attribute until
other attributes are desired to be defined. The "version" and
"eeprom_data_version" fields of both the template and the actual
definition must match, otherwise an error will be thrown.

If the definition file overrides all fields of the template file, the
template file is redundant. This results in an error.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>