fix: allow overriding eeprom_data_version field
TODO:
- fix format spec
- fix schema

In the original implementation of the templating functionality, the
"eeprom_data_version" was checked to be the same in both the base file
and the template. This is wrong, as the field communicates if the data
changed, not the version of the format itself.

Thus allow the "eeprom_data_version" field to be overridden by the base
file if a template is included.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>